+++ ts/packages/knowledgeProcessor/package.json
@@ -14,6 +14,7 @@
   "exports": {
     ".": "./dist/index.js"
   },
+  "types": "./dist/index.d.ts",
   "scripts": {
     "build": "npm run tsc",
     "postbuild": "copyfiles -u 1 \"src/**/*Schema.ts\" dist",
@@ -21,7 +22,9 @@
     "prettier": "prettier --check . --ignore-path ../../.prettierignore",
     "prettier:fix": "prettier --write . --ignore-path ../../.prettierignore",
     "test": "node --no-warnings --experimental-vm-modules ./node_modules/jest/bin/jest.js --testPathPattern=\".*\\.spec\\.js\"",
-    "tsc": "tsc -p src"
+    "test:debug": "node --inspect-brk --no-warnings --experimental-vm-modules ./node_modules/jest/bin/jest.js --testPathPattern=\".*\\.spec\\.js\"",
+    "test:full": "node --no-warnings --experimental-vm-modules ./node_modules/jest/bin/jest.js",
+    "tsc": "tsc -b"
   },
   "dependencies": {
     "aiclient": "workspace:*",
@@ -32,9 +35,10 @@
     "@types/node": "^18.18.7",
     "@types/jest": "^29.5.7",
     "copyfiles": "^2.4.1",
+    "dotenv": "^16.3.1",
+    "jest": "^29.7.0",
     "prettier": "^3.2.5",
     "rimraf": "^5.0.5",
-    "jest": "^29.7.0",
     "typescript": "^5.4.2"
   }
 }
