+++ ts/packages/knowledgeProcessor/tsconfig.json
@@ -1,0 +1,14 @@
+{
+  "extends": "../../tsconfig.base.json",
+  "compilerOptions": {
+    "composite": true,
+  },
+  "include": [],
+  "references": [
+    { "path": "./src" },
+    { "path": "./test" }
+  ],
+  "ts-node": {
+    "esm": true
+  }
+}
